requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
